initial commit for summaries refactor
