+++ mllib/src/main/scala/org/apache/spark/ml/clustering/KMeans.scala
@@ -18,13 +18,13 @@
 package org.apache.spark.ml.clustering
 
 import org.apache.hadoop.fs.Path
-
 import org.apache.spark.SparkException
 import org.apache.spark.annotation.{Experimental, Since}
 import org.apache.spark.ml.{Estimator, Model}
 import org.apache.spark.ml.linalg.{Vector, VectorUDT}
 import org.apache.spark.ml.param._
 import org.apache.spark.ml.param.shared._
+import org.apache.spark.ml.summary.{ClusteringSummary, KMeansSummary}
 import org.apache.spark.ml.util._
 import org.apache.spark.mllib.clustering.{KMeans => MLlibKMeans, KMeansModel => MLlibKMeansModel}
 import org.apache.spark.mllib.linalg.{Vector => OldVector, Vectors => OldVectors}
@@ -348,19 +348,4 @@ object KMeans extends DefaultParamsReadable[KMeans] {
   override def load(path: String): KMeans = super.load(path)
 }
 
-/**
- * :: Experimental ::
- * Summary of KMeans.
- *
- * @param predictions  `DataFrame` produced by `KMeansModel.transform()`.
- * @param predictionCol  Name for column of predicted clusters in `predictions`.
- * @param featuresCol  Name for column of features in `predictions`.
- * @param k  Number of clusters.
- */
-@Since("2.0.0")
-@Experimental
-class KMeansSummary private[clustering] (
-    predictions: DataFrame,
-    predictionCol: String,
-    featuresCol: String,
-    k: Int) extends ClusteringSummary(predictions, predictionCol, featuresCol, k)
+
