+++ mllib/src/test/java/org/apache/spark/ml/classification/JavaLogisticRegressionSuite.java
@@ -20,6 +20,7 @@ package org.apache.spark.ml.classification;
 import java.io.IOException;
 import java.util.List;
 
+import org.apache.spark.ml.summary.LogisticRegressionTrainingSummary;
 import org.junit.Assert;
 import org.junit.Test;
